moved src/bookmarkdialog.* to src/help/bookmarkdialog.*
